устранил предупреждения компилятора
4267: https://msdn.microsoft.com/ru-ru/library/6kck0s93.aspx
4244: https://msdn.microsoft.com/ru-ru/library/th7a07tz.aspx
4800: https://msdn.microsoft.com/ru-ru/library/b6801kcy.aspx

+++ src/bignum.h
@@ -188,7 +188,7 @@ public:
         pch[1] = (nSize >> 16) & 0xff;
         pch[2] = (nSize >> 8) & 0xff;
         pch[3] = (nSize) & 0xff;
-        BN_mpi2bn(pch, p - pch, this);
+        BN_mpi2bn(pch, (int)(p - pch), this);
     }
 
     uint64_t getuint64()
